// packages/app/src/hooks/useStats/types.ts
// Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
// SPDX-License-Identifier: GPL-3.0-only

export type StatKey =
	| 'averageRewardRate'
	| 'supplyStaked'
	| 'nextReward'
	| 'activeValidators'
	| 'averageCommission'
	| 'totalValidators'
	| 'minimumToJoinPool'
	| 'activePools'
	| 'minimumToCreatePool'
	| 'activeNominators'
	| 'minimumNominatorBond'
	| 'minimumActiveStake'
	| 'rewardCalculator'